Added file from Abandonware site: vetusware.com
Note I purchased the original program and have the disks and books.
As far as I can tell, the only purpose of this file would be to verify the file conversion. Certainly there is little reason to start using FirstChoice these days, despite it's utility. It doesn't handle modern file formats, and the data field sizes are small for uses such as images and video.